--- apps/aides/integration/__init__.py
@@ -1,3 +1,0 @@
-from ._grist import GristIntegration
-
-__all__ = ["GristIntegration"]
