+++ apps/aides/integration/eau_grandsudouest.py
@@ -1,9 +1,8 @@
-import functools
 import logging
 
 from markdownify import markdownify as md
 
-from ._grist import AbstractRawFields, AbstractAidesSource
+from ._base import AbstractRawFields, AbstractAidesSource
 from ._utils import get_soup_from_url
 
 logger = logging.getLogger(__name__)
@@ -53,26 +52,3 @@ class EauGrandSudOuest(AbstractAidesSource):
                 aides.append(data)
 
         return aides
-
-    @functools.cached_property
-    def _organisme(self) -> tuple[str, str]:
-        return self.grist_integration.build_grist_organisme(
-            "Agence de l'eau Grand Sud-Ouest"
-        )
-
-    @functools.cached_property
-    def _sujets(self) -> tuple[str, ...]:
-        return self.grist_integration.build_grist_sujets(["Eau"])
-
-    def _enrich_aide(self, aide: dict) -> None:
-        columns = self.grist_integration.VisibleSolutionsColumns
-        aide.update(
-            {
-                columns.ORGANISME.value: self._organisme,
-                columns.SUJETS.value: self._sujets,
-                columns.NOM.value: aide[Fields.NOM.name_full],
-                columns.URL_DESCRIPTIF.value: aide[Fields.URL.name_full],
-                columns.DESCRIPTION.value: aide[Fields.DESCRIPTION.name_full],
-                columns.MONTANT_TAUX.value: aide[Fields.OBJECTIFS_ET_TAUX.name_full],
-            }
-        )
