+++ apps/aides/integration/eau_loire_bretagne.py
@@ -1,8 +1,7 @@
-import functools
 import logging
 from collections import defaultdict
 
-from ._grist import AbstractAidesSource, AbstractRawFields
+from ._base import AbstractAidesSource, AbstractRawFields
 from ._utils import get_soup_from_url
 
 logger = logging.getLogger(__name__)
@@ -64,26 +63,3 @@ class EauLoireBretagne(AbstractAidesSource):
                         el.get_text(strip=True) + "\n\n"
                     )
         return to_load
-
-    @functools.cached_property
-    def _organisme(self) -> tuple[str, str]:
-        return self.grist_integration.build_grist_organisme(
-            "Agence de l'eau Loire-Bretagne"
-        )
-
-    def _enrich_aide(self, aide: dict):
-        columns = self.grist_integration.__class__.VisibleSolutionsColumns
-        aide.update(
-            {
-                columns.ORGANISME.value: self._organisme,
-                columns.NOM.value: aide[Fields.NOM.name_full],
-                columns.PROMESSE.value: aide[Fields.ENJEU.name_full]
-                + ", "
-                + aide[Fields.ACTION.name_full],
-                columns.URL_DESCRIPTIF.value: aide[Fields.URL.name_full],
-                columns.MONTANT_TAUX.value: aide[Fields.TAUX.name_full],
-                columns.CONDITIONS.value: aide[Fields.CONDITIONS.name_full],
-                columns.TYPE_DEPENSE.value: aide[Fields.COMPLEMENTS.name_full],
-                columns.AUTRES_INFORMATIONS.value: aide[Fields.NOTES.name_full],
-            }
-        )
