+++ pyproject.toml
@@ -9,7 +9,6 @@ dependencies = [
   "django-anymail[brevo] ~=13.0",
   "django-csp ~=4.0",
   "django-dsfr ~=2.1",
-  "django-grist-loader ~=0.3",
   "django-pgtrigger ~=4.15",
   "django-prose ~=2.1",
   "django-reversion ~=5.1",
@@ -20,9 +19,9 @@ dependencies = [
   "markdown ~=3.8",
   "markdownify ~=1.1",
   "mjml-python ~=1.3",
-  "qrcode ~=7.4",
   "psycopg[binary] ~=3.0",
-  "mjml-python ~=1.3",
+  "pygrister ~=0.7",
+  "qrcode ~=7.4",
   "requests ~=2.32",
   "sentry-sdk[django] ~=2.26",
   "whitenoise ~=6.8",
